Fix weird edge case with ambiguous parse
In some quickcheck tests I noticed that the schedule * * 31 9 * was
being checked as * * * 9 * and passing when it shouldn't. Turns out this
was due to quickcheck exploring some ambiguous cases unlikely to be hit
upon by users, such as a singleton list with star in it. We incorrectly
detected that as "restricted" and applied the special case day of week
OR day of month trick.