Bridge larger gap of quickcheck versions
quickcheck does some CPP and will conditionally define a nonempty
instance for arbitrary when possible. This means our test suite either
fails on some new versions for duplicate instance definitions for
nonempty or fails on older versions for it being missing. We don't use
nonempty pervasively so its easier to do without it.